refactor: deriv-ez svg imports
