+++ src/components/elements/off-canvas-menu.tsx
@@ -31,7 +31,7 @@ import API from 'images/svg/menu/developers.svg'
 import Diagonal from 'images/svg/elements/pink-right-diagonal.svg'
 import DMT5 from 'images/svg/dmt5/dmt5-icon.svg'
 import DerivX from 'images/svg/custom/deriv-x.svg'
-import DerivEZ from 'images/svg/deriv-ez/deriv-ez-logo-red.svg'
+import { DerivEzLogoRed } from 'images/svg/deriv-ez'
 import DTrader from 'images/svg/dtrader/dtrader-icon.svg'
 import DerivGo from 'images/svg/deriv-go/deriv-go-icon.svg'
 import Forex from 'images/svg/custom/forex-nav.svg'
@@ -282,7 +282,7 @@ export const OffCanvasMenuWrapper = (props: OffCanvasMenuWrapperProps) => {
                                         aria_label="Derivez"
                                         Icon={() => (
                                             <img
-                                                src={DerivEZ}
+                                                src={DerivEzLogoRed}
                                                 alt={localize('_t_Deriv EZ trading paltform_t_')}
                                                 width="32"
                                                 height="32"
