+++ src/features/pages/deriv-ez/deprecated-other-platforms/index.tsx
@@ -1,9 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import dxIcon from 'images/svg/deriv-ez/dx-icon.svg'
-import dtIcon from 'images/svg/deriv-ez/dt-icon.svg'
-import dmt5Icon from 'images/svg/deriv-ez/dmt5-icon.svg'
-import dGoIcon from 'images/svg/deriv-ez/go-icon.svg'
-import dbIcon from 'images/svg/deriv-ez/db-icon.svg'
+import { DbIcon, DGoIcon, Dmt5Icon, DtIcon, DxIcon } from 'images/svg/deriv-ez'
 import GenericCarousel, { TRenderableData } from 'components/custom/carousel/_platforms-carousel'
 
 const OurPlatforms = () => {
@@ -16,31 +12,31 @@ const OurPlatforms = () => {
 
     const carouselData: TRenderableData[] = [
         {
-            icon: dmt5Icon,
+            icon: Dmt5Icon,
             heading: 'Deriv MT5',
             paragraph: '_t_The all-in-one CFD trading platform._t_',
             link: '/dmt5/',
         },
         {
-            icon: dtIcon,
+            icon: DtIcon,
             heading: 'Deriv Trader',
             paragraph: '_t_Our flagship app for trading options and multipliers._t_',
             link: '/dtrader/',
         },
         {
-            icon: dGoIcon,
+            icon: DGoIcon,
             heading: 'Deriv Go',
             paragraph: '_t_Our best trading experience on your mobile._t_',
             link: '/deriv-go/',
         },
         {
-            icon: dxIcon,
+            icon: DxIcon,
             heading: 'Deriv X',
             paragraph: '_t_The CFD trading platform to fit your style._t_',
             link: '/derivx/',
         },
         {
-            icon: dbIcon,
+            icon: DbIcon,
             heading: 'Deriv Bot',
             paragraph: '_t_Automate your trading. No coding required._t_',
             link: '/dbot/',
