+++ src/features/pages/deriv-ez/why-trade/data.tsx
@@ -1,8 +1,6 @@
 import React from 'react'
 import { DerivEzCardTypes } from '../types'
-import AnyTime from 'images/svg/deriv-ez/twenty-four-seven.svg'
-import FavouriteAssets from 'images/svg/deriv-ez/favourite-assets.svg'
-import EasyStart from 'images/svg/deriv-ez/easy-to-start.svg'
+import { EasyToStart, FavouriteAssets, TwentyFourSeven } from 'images/svg/deriv-ez'
 import Image from 'features/components/atoms/image'
 
 export const card_data: DerivEzCardTypes[] = [
@@ -10,13 +8,13 @@ export const card_data: DerivEzCardTypes[] = [
         title: '_t_Easy to start_t_',
         subtitle:
             '_t_Start trading CFDs in two easy steps — click to get an account, start trading. No need to remember another login ID and password._t_',
-        icon: <Image src={EasyStart} height={48} alt={'_t_Easy start_t_'} />,
+        icon: <Image src={EasyToStart} height={48} alt={'_t_Easy start_t_'} />,
     },
     {
         title: '_t_24/7 trading_t_',
         subtitle:
             '_t_Trade anytime you want, including weekends and holidays, on cryptocurrencies and our proprietary derived indices._t_',
-        icon: <Image src={AnyTime} height={48} alt={'_t_24/7 chat_t_'} />,
+        icon: <Image src={TwentyFourSeven} height={48} alt={'_t_24/7 chat_t_'} />,
     },
     {
         title: '_t_All your favourite assets, one platform_t_',
